dragon-senate-saas/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langgraph==0.2.39
langchain==0.3.1
langgraph-checkpoint-postgres==2.0.2
psycopg[binary,pool]==3.2.2
asyncpg==0.29.0
pydantic==2.9.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
